Add TRADING_MODE flag, delete migration scripts, remove dead code
Introduces TRADING_MODE env var (LIVE/OFF) to gate all order placement
for a training/observation deployment. Deletes 14 migration scripts
(~2,300 lines), removes inline migration code from brier_scoring,
state_manager, and utils, and removes the never-read
compliance_decisions.csv sink.

- TRADING_MODE=OFF gates: place_order, place_directional_spread_with_protection,
  place_queued_orders, close_stale_positions, emergency_hard_close,
  close_stale_positions_fallback, run_position_audit_cycle, cancel_all_open_orders,
  close_spread_with_protection_cleanup
- Notifications prefixed with OFF, Cockpit shows warning banner
- Extracted resolve_with_cycle_aware_match to trading_bot/brier_reconciliation.py
- Removed migrate_legacy_entries from state_manager and its orchestrator call
- Removed compliance_decisions.csv log_compliance_decision + 2 call sites
- Removed migration step from deploy.sh
- 261 tests pass, 0 failures
