+++ notifications.py
@@ -43,7 +43,11 @@ def send_pushover_notification(config: dict, title: str, message: str, attachmen
 
     # Load the environment name, defaulting to nothing if not set
     env_prefix = os.getenv("ENV_NAME", "")
-    
+
+    # Prepend OFF indicator if trading is disabled
+    if os.getenv("TRADING_MODE", "LIVE").upper().strip() == "OFF":
+        env_prefix = f"OFF {env_prefix}".strip() if env_prefix else "OFF"
+
     # Only add the prefix if it exists
     if env_prefix:
         title = f"{env_prefix} - {title}"
